requirements.txt
# Core API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0

# AI - LangGraph + Gemini (recommended)
langgraph>=0.0.40
google-generativeai>=0.3.0

# AI - Alternative providers (optional)
#openai>=1.0.0
#anthropic>=0.7.0
#httpx>=0.25.0  # For Ollama

# File handling
python-multipart>=0.0.6
lxml>=4.9.0

# Testing
pytest>=9.0.0

# Demo CLI
requests>=2.31.0
rich>=13.0.0
